pyproject.toml
[project]
name = "research-workspace"
version = "0.1.0"
description = "Research workspace for experiments"
requires-python = ">=3.10"
dependencies = [
    "datasets>=4.4.2",
    "matplotlib>=3.10.8",
    "numpy>=2.2.6",
    "openai>=2.14.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.7.2",
    "scipy>=1.15.3",
    "seaborn>=0.13.2",
    "tenacity>=9.1.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/research_workspace"]
